Added seperate script for tflite conversion and prediction
